+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.1.1 [2015-11-15]
 
-- Added ability ordering of visible a columns.
+- Added ability sortering of visible a columns.
 
 ## v1.1.0 [2015-11-13]